# Reduce LogEventLevel boxing