Hi!

**What issue does this PR address?**

During the profiling I found that `LogEventLevel` was boxed during [formatting](https://github.com/serilog/serilog/blob/dev/src/Serilog/Formatting/Display/LevelOutputFormat.cs#L55). You can see the evidence of the boxing in the screenshot :

<img width="1211" alt="logeventlevel boxing" src="https://user-images.githubusercontent.com/664889/49733899-8e7a0e80-fca4-11e8-835d-905411971210.png">

**Does this PR introduce a breaking change?**
Nope

**Please check if the PR fulfills these requirements**
- [X] The commit follows our [guidelines](https://github.com/serilog/serilog/blob/dev/CONTRIBUTING.md)
- [ ] Unit Tests for the changes have been added (for bug fixes / features)

**Other information**: