@Pliner 2019-01-14
Here are two benchmark results: the baseline(commit hash 5c3a782) and the PR.

As far as I understand cases of the benchmark, the most relevant is `OutputTemplateRenderingBenchmark`.

# net46
``` ini

BenchmarkDotNet=v0.10.6, OS=Windows 10 Redstone 2 (10.0.15063)
Processor=Intel Core i7-6700 CPU 3.40GHz (Skylake), ProcessorCount=8
Frequency=3328125 Hz, Resolution=300.4695 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.7.3190.0
  DefaultJob : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.7.3190.0


```
 |         Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
 |--------------- |---------:|----------:|----------:|-------:|----------:|
 | FormatToOutput | 2.280 us | 0.0203 us | 0.0180 us | 0.1144 |     492 B |

 |         Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
 |--------------- |---------:|----------:|----------:|-------:|----------:|
 | FormatToOutput | 1.729 us | 0.0077 us | 0.0060 us | 0.1068 |     456 B |

# netcoreapp2.0

``` ini

BenchmarkDotNet=v0.10.6, OS=Windows 10 Redstone 2 (10.0.15063)
Processor=Intel Core i7-6700 CPU 3.40GHz (Skylake), ProcessorCount=8
Frequency=3328125 Hz, Resolution=300.4695 ns, Timer=TSC
dotnet cli version=2.1.4
  [Host]     : .NET Core 4.6.26614.01, 64bit RyuJIT
  DefaultJob : .NET Core 4.6.26614.01, 64bit RyuJIT


```
 |         Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
 |--------------- |---------:|----------:|----------:|-------:|----------:|
 | FormatToOutput | 1.315 us | 0.0236 us | 0.0221 us | 0.1163 |     488 B |

 |         Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
 |--------------- |---------:|----------:|----------:|-------:|----------:|
 | FormatToOutput | 1.221 us | 0.0116 us | 0.0109 us | 0.1087 |     464 B |

It seems there is no perfomance degradation 😅

PS Also here are full results.

[baseline.zip](https://github.com/serilog/serilog/files/2754037/baseline.zip)
[pr.zip](https://github.com/serilog/serilog/files/2754036/pr.zip)