@Pliner 2019-01-16
@nblumhardt Is it possible to include the PR to 2.8 release?