@nblumhardt 2019-01-16
Thanks for the nudge @Pliner - sorry about the slow turnaround on these. Optimisations especially need a close look and I am still a little way from having time to dig into these in detail. Should be low risk to roll them into a 2.8.1, though!